# Improve performance by relying on a native string instead of InputStream