… test/HTML5/Parser/FileInputStreamTest.html
… test/HTML5/Parser/FileInputStreamTest.php
… test/HTML5/Parser/StringInputStreamTest.php