When I used the library in the context of https://github.com/tgalopin/html-sanitizer, I stumbled upon speed issues. I launched a Blackfire analysis and saw it mainly came from this library.

I started to investigate and quickly found out that the architecture with an InputStream and a Scanner was the biggest part of the problem: calling 3 methods for each characters of the input slowed a lot the parsing process.

In addition to that, I realized that the InputStream architecture, while being very extensible, is not used in the library itself: the FileStreamInput is a plain StringInput with a `file_get_contents`, and the StringInput is a simple wrapper around a native string. 

Finally, I think it is not likely that anyone using this library did it in a way where they can't transform their input as a string.

Thus, I would like to propose in this pull request to simplify the library critical execution path by deprecating (and removing in 3.0) the InputStream architecture and relying on a native string directly in the Scanner instead.

This implementation shows great speed improvements:

**Test script**

```php
require __DIR__.'/../vendor/autoload.php';

$html = file_get_contents(__DIR__.'/fixture.html');

$start = microtime(true);
$total = 100;

for ($i = 0; $i < $total; $i++) {
    $parser = new \Masterminds\HTML5();
    $parser->loadHTMLFragment($html);
}

$time = microtime(true) - $start;

echo 'Total time: '.round($time * 1000)."ms\n";
echo 'Time per iteration: '.round(($time * 1000) / $total, 2)."ms\n";
```

**Simple time measurement**

*Before*:

```
Total time: 935ms
Time per iteration: 9.35ms
```

*After*:

```
Total time: 731ms
Time per iteration: 7.31ms
```

**Blackfire analysis**

> Note that Blackfire introduces an overhead on functions and methods calls, which enhance the difference. This is why I also did simple time measurements.

- *Before*: 245ms (https://blackfire.io/profiles/c4ff741d-a470-4f91-8553-e37445b3ce17/graph)
- *After*: 144ms (https://blackfire.io/profiles/4e1a4743-f72b-409c-8ab6-767ff9fb3181/graph)


I am fully aware this could be a bit controversial PR but I really think it would be a great improvement for this library, especially in the context of Drupal. In many cases, we can expect a 22% speed improvement, which is worth the removal of InputStream in my opinion :) !