@tgalopin 2018-11-04
I just pushed an additional performance gain achieved by removing some calls to the `Scanner::current()` method in the Tokenizer.

I updated the script to run 1000 iterations in order to get more precise results. Here are the results for the whole PR now:

*Before:*

```
Total time: 9361ms
Time per iteration: 9.36ms
```

*After removing the InputStream system:*

```
Total time: 7559ms
Time per iteration: 7.56ms
```

*After (whole PR):*

```
Total time: 7389ms
Time per iteration: 7.39ms
```