@PhenX 2018-11-04
@tgalopin while reducing parsing time is very good, did you measure memory usage too? I think streams were used in order to reduce memory footprint.