@tgalopin 2018-11-04
Thanks for the feedback :) !

InputStream instances were not actually streams, they were just wrappers around a plain string :) . The memory footprint is even lower in the new version (1.25 MB to 1.17MB, as you can see in the Blackfire profiles), probably because of the removal of the objects :) .

I do think though that using real streams may be a great idea for further improvements. I'll perhaps have a look if I find how to do it properly :) .