@tgalopin 2018-11-04
I removed the tokenizer improvements from this PR to focus on the InputStream here. I'm still working on the tokenizer, I'll open a new PR once ready. Note that the two improvements are fully independent, no need to wait for the other if we want to merge this one :) .