@tgalopin 2018-11-05
Here it is: https://github.com/Masterminds/html5-php/pull/147 :) !